InsightMate/Scripts/requirements.txt
flask
openai
google-auth
google-auth-oauthlib
google-api-python-client
requests
msal
PyPDF2
python-docx
textract
dateparser

